+++ uguisu-engine/src/analyze.rs
